requirements.txt
Django==1.10
djangorestframework==3.6.4
pytz==2017.2
